Fix stuffs

+++ trainer/models/lumina2_loader.py
@@ -20,7 +20,7 @@ def load_lumina2_components(args, device=None, weight_dtype=torch.float32):
             args.pretrained_model_name_or_path, subfolder="scheduler",
         )
     else:
-        # Default config if not loading from pretrained (though usually we do)
+        # Default config
         noise_scheduler = FlowMatchEulerDiscreteScheduler()
 
     # Load Tokenizer
@@ -31,10 +31,9 @@ def load_lumina2_components(args, device=None, weight_dtype=torch.float32):
     )
 
     # Load Text Encoder (Gemma 2)
-    # Gemma 2 is often loaded in bfloat16 or float16 to save memory
     text_encoder_dtype = weight_dtype
     if weight_dtype == torch.float16 or weight_dtype == torch.bfloat16:
-        text_encoder_dtype = weight_dtype # Keep consistent
+        text_encoder_dtype = weight_dtype 
     
     text_encoder = AutoModelForCausalLM.from_pretrained(
         args.pretrained_model_name_or_path, 
@@ -46,17 +45,15 @@ def load_lumina2_components(args, device=None, weight_dtype=torch.float32):
     )
 
     # Load VAE (AutoencoderKL)
-    # VAE is usually kept in float32 for stability, but we can respect weight_dtype if forced
-    # The provided script keeps VAE in float32.
     vae = AutoencoderKL.from_pretrained(
         args.vae_path,
         device_map=device_map,
-        torch_dtype=torch.float32, 
+        torch_dtype=torch.bfloat16,
         revision=getattr(args, "revision", None),
         variant=getattr(args, "variant", None),
     )
 
-    # Load Transformer (Lumina2Transformer2DModel)
+    # Load Transformer
     model_config = getattr(args, "model_config", None)
     if model_config:
         print(f"Initializing Lumina2Transformer2DModel from config: {model_config}")
